更新fork项目

+++ apps/microtvm/reference-vm/zephyr/base-box/setup.sh
@@ -18,6 +18,13 @@
 
 set -e
 
+skip_zeroing_disk=0
+if [ -e "$HOME/skip_zeroing_disk" ]; then
+    echo "NOTE: will not zero disk at the end due to VMWare Fusion bug"
+    echo "See: https://communities.vmware.com/t5/VMware-Fusion-Discussions/VMWare-Fusion-Pro-11-15-6-16696540-causes-macOS-crash-during/m-p/2284011#M139190"
+    skip_zeroing_disk=1
+fi
+
 sudo apt update
 sudo apt install -y build-essential
 sudo apt-get --purge remove modemmanager  # required to access serial ports.
@@ -52,34 +59,40 @@ sudo apt install -y llvm
 sudo apt install -y protobuf-compiler libprotoc-dev
 
 # nrfjprog
+NRF_COMMANDLINE_TOOLS_FILE=nRFCommandLineToolsLinuxamd64.tar.gz
+NRF_COMMANDLINE_TOOLS_URL=https://www.nordicsemi.com/-/media/Software-and-other-downloads/Desktop-software/nRF-command-line-tools/sw/Versions-10-x-x/10-12-1/nRFCommandLineTools10121Linuxamd64.tar.gz
+NRF_COMMANDLINE_TOOLS_INSTALLER=nRF-Command-Line-Tools_10_12_1_Linux-amd64.deb
+JLINK_LINUX_INSTALLER=JLink_Linux_V688a_x86_64.deb
+
 cd ~
 mkdir -p nrfjprog
-wget --no-verbose -O nRFCommandLineTools1090Linuxamd64.tar.gz https://www.nordicsemi.com/-/media/Software-and-other-downloads/Desktop-software/nRF-command-line-tools/sw/Versions-10-x-x/10-9-0/nRFCommandLineTools1090Linuxamd64tar.gz
+wget --no-verbose -O $NRF_COMMANDLINE_TOOLS_FILE $NRF_COMMANDLINE_TOOLS_URL
 cd nrfjprog
-tar -xzvf ../nRFCommandLineTools1090Linuxamd64.tar.gz
-sudo apt install -y ./JLink_Linux_V680a_x86_64.deb
-sudo apt install -y ./nRF-Command-Line-Tools_10_9_0_Linux-amd64.deb
+tar -xzvf "../${NRF_COMMANDLINE_TOOLS_FILE}"
+sudo apt install -y "./${JLINK_LINUX_INSTALLER}"
+sudo apt install -y "./${NRF_COMMANDLINE_TOOLS_INSTALLER}"
 source ~/.profile
 nrfjprog --help
 cd ..
-rm -rf nrfjprog nRFCommandLineTools1090Linuxamd64.tar.gz
+rm -rf nrfjprog "${NRF_COMMANDLINE_TOOLS_FILE}"
 
 # Zephyr
 pip3 install --user -U west
 echo 'export PATH=$HOME/.local/bin:"$PATH"' >> ~/.profile
 source ~/.profile
 echo PATH=$PATH
-west init --mr v2.4.0 ~/zephyr
-cd ~/zephyr
-west update
-west zephyr-export
+REPO_ROOT=$(git rev-parse --show-toplevel)
+${REPO_ROOT}/docker/install/ubuntu_init_zephyr_project.sh ~/zephyr v2.5.0
 
 cd ~
 echo "Downloading zephyr SDK..."
-wget --no-verbose https://github.com/zephyrproject-rtos/sdk-ng/releases/download/v0.11.3/zephyr-sdk-0.11.3-setup.run
-chmod +x zephyr-sdk-0.11.3-setup.run
-./zephyr-sdk-0.11.3-setup.run -- -d ~/zephyr-sdk -y
-rm -rf zephyr-sdk-0.11.3-setup.run
+ZEPHYR_SDK_VERSION=0.12.3
+ZEPHYR_SDK_FILE=zephyr-sdk-linux-setup.run
+wget --no-verbose -O $ZEPHYR_SDK_FILE \
+    https://github.com/zephyrproject-rtos/sdk-ng/releases/download/v${ZEPHYR_SDK_VERSION}/zephyr-sdk-${ZEPHYR_SDK_VERSION}-x86_64-linux-setup.run
+chmod +x $ZEPHYR_SDK_FILE
+"./$ZEPHYR_SDK_FILE" -- -d ~/zephyr-sdk -y
+rm -rf "${ZEPHYR_SDK_FILE}"
 
 # GDB for Zephyr SDK depends on python3.8
 sudo add-apt-repository ppa:deadsnakes/ppa
@@ -96,10 +109,15 @@ sed -i "/^# If not running interactively,/ i\\ " ~/.bashrc
 
 # Clean box for packaging as a base box
 sudo apt-get clean
-EMPTY_FILE="$HOME/EMPTY"
-dd if=/dev/zero "of=${EMPTY_FILE}" bs=1M || /bin/true
-if [ ! -e "${EMPTY_FILE}" ]; then
-    echo "failed to zero empty sectors on disk"
-    exit 2
+if [ $skip_zeroing_disk -eq 0 ]; then
+    echo "Zeroing disk..."
+    EMPTY_FILE="$HOME/EMPTY"
+    dd if=/dev/zero "of=${EMPTY_FILE}" bs=1M || /bin/true
+    if [ ! -e "${EMPTY_FILE}" ]; then
+        echo "failed to zero empty sectors on disk"
+        exit 2
+    fi
+    rm -f "${EMPTY_FILE}"
+else
+    echo "NOTE: skipping zeroing disk due to command-line argument."
 fi
-rm -f "${EMPTY_FILE}"
